Move synthetic admixture dataset utilities to snputils.datasets
