+++ snputils/datasets/__init__.py
@@ -2,6 +2,14 @@ from .load_dataset import (
     available_datasets_list,
     load_dataset,
 )
+from .synthetic_admixture import (
+    DEFAULT_ADMIXTURE_ANCESTRY_MAP,
+    build_feature_layout,
+    build_lai_object,
+    build_synthetic_admixture_dataset,
+    covariate_coefficients,
+    standardize,
+)
 from ._registry import (
     ChromosomeResource,
     DatasetSpec,
@@ -13,7 +21,13 @@ __all__ = [
     "ChromosomeResource",
     "DatasetSpec",
     "PopulationMetadataSpec",
+    "DEFAULT_ADMIXTURE_ANCESTRY_MAP",
     "available_datasets_list",
+    "build_feature_layout",
+    "build_lai_object",
+    "build_synthetic_admixture_dataset",
+    "covariate_coefficients",
     "load_dataset",
     "register_dataset",
+    "standardize",
 ]
